# Alpha complex is using 3d version in dim 3