* Factory to build python alpha complex.
* 3d version of alpha complex.
* Needs change in doc as alphacomplexdoc.off is a fake 3d off data file