@mglisse 2020-06-28
I also pushed a few clean-ups that I did while looking at the branch.