@VincentRouvreau 2020-07-02
> Assuming you checked that it gives a performance gain in 3D and doesn't hurt in some other dimension, ok.

Before the modifications:
| dimension | 1000 pts | 2000 pts |
| --- | --- | --- |
| 2 | 1.1 | 1.3 |
| 3 | 1.4 | 1.9 |
| 4 | 7. | 12.6 |

After the modifications:
| dimension | 1000 pts | 2000 pts |
| --- | --- | --- |
| 2 | 1.2 | 1.2 |
| 3 | 1.2 | 1.3 |
| 4 | 6.9 | 12.8 |

Seems good to me (times are user+sys when using bash time utils on alpha_complex and simplex tree creation).