@mglisse 2020-07-02
Uh, are you saying that computing the alpha-complex on 2000 points takes exactly the same time as for 1000 points in dimension 2? That looks suspicious...