@VincentRouvreau 2020-07-02
I did some more timings in dimension 2 with the new algo:
| 1000 pts | 2000 pts |
| --- | --- |
| 1.217 | 1.247 |
| 1.189 | 1.341 |
| 1.053 | 1.396 |
| 1.202 | 1.339 |
| 1.145 | 1.292 |

Launching Python is about 1.1 sec.