@mglisse 2020-07-02
You would get much more accurate timings running this in ipython
```python
import gudhi
import numpy as np
import timeit
a=np.random.rand(100,2)
%timeit gudhi.AlphaComplex(points=a).create_simplex_tree()
```
It does seem to give a (surprisingly) small performance gain in 3D and make no difference in dimension 2 or 4, ok.