@VincentRouvreau 2020-07-02
Ok, so I changed the way it was benchmarked using:
```python
import gudhi                                                            
def alpha(off): 
    alpha_complex = gudhi.AlphaComplex(off_file=off) 
    simplex_tree = alpha_complex.create_simplex_tree() 
    print(simplex_tree.num_simplices()) 

time alpha('/my/path/to/file.off')
```

Before the modifications:
dimension | 1000 pts | 2000 pts
-- | -- | --
2 | 0.036 | 0.103
3 | 0.292 | 0.738
4 | 5.75 | 11.3

After the modifications:
dimension | 1000 pts | 2000 pts
-- | -- | --
2 | 0.062 | 0.126
3 | 0.108 | 0.193
4 | 5.95 | 11.6