@mglisse 2020-07-02
I wonder why you are getting this great speed-up (3x) while I only get a modest speed-up (from 90ms to 70ms for 1000 points), even using your script (I had to add `%` in front of `time` to make it work in ipython), with a file containing 1000 random points generated uniformly in a cube. What kind of data are you testing on? What version of cgal? (I am using 5.0.2) Are you comparing to master just before the merge?